__int64 zu long long int gemacht

+++ src/elliptic_curve_factorization.h
@@ -1,6 +1,6 @@
 #ifndef ELLIPTIC_CURVE_FACTORIZATION_H
 #define ELLIPTIC_CURVE_FACTORIZATION_H
 
-void ecm_factorization(__int64 n, __int64* p, __int64* q);
+void ecm_factorization(long long int n, long long int* p, long long int* q);
 
 #endif
