+++ src/factorization.cu
@@ -1,7 +1,7 @@
 #include "factorization.h"
 #include "stdio.h"
 
-void factorization(__int64 n, __int64* p, __int64* q) {
+void factorization(long long int n, long long int* p, long long int* q) {
 	int i , anz =0;
 	for ( i =2; i <= n ; i++)
 		while ( n % i == 0)
