+++ src/gpu_factorization.cu
@@ -1,13 +1,13 @@
 #include "gpu_factorization.h"
 
-__global__ void gpu_factorization(__int64* n, __int64* p, __int64* q) {
+__global__ void gpu_factorization(long long int* n, long long int* p, long long int* q) {
 
 
-	__int64 local_n = *n;
-	__int64 i;
-	__int64 idx = blockIdx.x * blockDim.x + threadIdx.x;
-	__int64 step_size = blockDim.x*gridDim.x;
-	__int64 steps = local_n/2;
+	long long int local_n = *n;
+	long long int i;
+	long long int idx = blockIdx.x * blockDim.x + threadIdx.x;
+	long long int step_size = blockDim.x*gridDim.x;
+	long long int steps = local_n/2;
 	//TODO: n = sqrt(n) !
 
 	for (i=3+idx; i<steps; i += step_size)
